requirements.txt
streamlit==1.20.0
pandas
pillow==8.4.0
python-dotenv==0.19.0
requests==2.26.0
gunicorn==20.1.0 